+++ src/registry/domain/components-cache/index.js
@@ -2,7 +2,6 @@
 
 const _ = require('underscore');
 
-const ComponentsDetails = require('./components-details');
 const ComponentsList = require('./components-list');
 const eventsHandler = require('../events-handler');
 const getUnixUTCTimestamp = require('../../../utils/get-unix-utc-timestamp');
@@ -12,7 +11,6 @@ module.exports = function(conf, cdn){
   let cachedComponentsList, refreshLoop;
 
   const componentsList = ComponentsList(conf, cdn);
-  const componentsDetails = ComponentsDetails(conf, cdn);
 
   const poll = () => setTimeout(() => {
     componentsList.getFromJson((err, data) => {
@@ -49,32 +47,7 @@ module.exports = function(conf, cdn){
       callback(null, cachedComponentsList);
     },
 
-    getDetails: (callback) => componentsDetails.getFromJson(callback),
-    
     load: (callback) => {
-
-      const verifyDetailsIntegrity = (componentsList) => {
-        const next = () => cacheDataAndStartPolling(componentsList, callback);
-
-        componentsDetails.getFromJson((jsonErr, jsonDetails) => {
-          componentsDetails.getFromDirectories(componentsList.components, jsonDetails, (dirErr, dirDetails) => {
-
-            if(dirErr){
-              return returnError('components_details_get', dirErr, callback);
-            } else if(jsonErr || !_.isEqual(dirDetails.components, jsonDetails.components)){
-              componentsDetails.save(dirDetails, (saveErr) => {
-                if(saveErr){
-                  return returnError('components_details_save', saveErr, callback);
-                }
-
-                next();
-              });
-            } else {
-              next();
-            }
-          });
-        });
-      };
 
       componentsList.getFromJson((jsonErr, jsonComponents) => {
         componentsList.getFromDirectories((dirErr, dirComponents) => {
@@ -85,10 +58,10 @@ module.exports = function(conf, cdn){
               if(saveErr){
                 return returnError('components_list_save', saveErr, callback);
               }
-              verifyDetailsIntegrity(dirComponents);
+              cacheDataAndStartPolling(dirComponents, callback);
             });
           } else {
-            verifyDetailsIntegrity(jsonComponents);
+            cacheDataAndStartPolling(jsonComponents, callback);
           }
         });
       });
@@ -99,6 +72,7 @@ module.exports = function(conf, cdn){
         if(err){
           return returnError('components_cache_refresh', err, callback);
         }
+
         cacheDataAndStartPolling(components, callback);
       });
     }
